pyproject.toml
[project]
name = "integration-test-template"
version = "0.1.0"
description = "Integration test"
readme = "README.md"

requires-python = ">=3.12"
dependencies = [
    # OBS: databricks-connect package *replaces* pyspark. Do not install both at the same time.
    #"pyspark>=4.0.0,<5.0.0",
]

[dependency-groups]
unit-tests = [
    # Thise will break integration test relying on databricks-connect
    #"pyspark>=4.0.0,<5.0.0",
]
integration-tests = [
    "databricks-connect==17.1.0",
    "pytest>=8.3.4",
]

[tool.uv]
default-groups = ["unit-tests"]

# [project.scripts]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ps_test_blueprint"]
